Master (#2)
* This definitely isn't needed anymore

* //TODO complete

* Turtle-Service Light Housekeeping (#574)

* Removes artifacts from TRTL Forks 

When people fork TRTL and dont do the whole job, our network tends to WTF seeing people with different genesis blocks connecting.

* Removing more boost

* [ENHANCEMENT] Add CN Soft Shell Support (#569)

* CN Slow Hash

* Creating standard tests

* Missed this in the portable Crypto routine

* Bumping version on development branch after release of 0.8.4

* [WIP] LWMA-3 (#571)

LWMA-3

* Bumping P2P Current Version for 0.10.0 -- will need to bump minimum version to `3` closer to 0.10.0 release

This is required due to the network upgrade (soft-fork) to LWMA3 at block 1M

* Travis Fix: (ノಠ益ಠ)ノ彡┻━┻

* [ENHANCEMENT] Added Additional Performance Benchmarks to CryptoTest (#577)

* Adding additional performance tests for v1/v2 as there is a difference

* Need to make sure that we check our data size before proceeding into the variants

* Dear Boost,

We think we need to spend some time to think about where this relationship is going. We've grown apart of over the years and we are not quite sure that this was meant to be.

While we're separated, feel free to see other people, other groups, different cryptocurrencies. We'll certainly play with other libraries that will very likely replace you in the end. There's still a chance though that we will come running back to you so definitely keep your hopes up. If you happen to find someone else that loves you as much as you need loved, go, be free.

For now, you can keep the boxes in the closet but at some point you're going to need to come and get them.

Sincerely,

The TurtleCoin Developers

* Requested changes

* Tweaks to makechange.py script

walletd to turtle-service
less loops
more python, needs python 3.6 now for secrets package
secrets used to get variety in random numbers thx to IBMC
  if min is 1 and max is 5000, numbers can actually be 2, 20, 200, 2,000 instead of all being thousands
made things easier for forks by moving params to the top
prints amounts out with messages for easy monitoring

* Update to make sure that version.h is a dependency of the necessary libraries

* [WIP] C++17 Support (#535)

* C++17 Support

* C++17 compilation instructions

* These can't be implict_values

* turtle-service: support old flat ini config for backward compatibility (#581)

* turtle-service: support old flat ini config for backward compatibility

* that's a big buffer (#585)

* that's a big buffer

* Update parsers to do megabytes -> bytes

* add block 2 checkpoint to assist people forking

If people skip the step in the guide regarding the checkpoints.
Having this block 2 checkpoint fail and spit out messages will
hopefully, prompt them to go read the guide again.

* Config Files Format Conversion from INI to JSON + other fixes (#588)

* config file update/conversion for smooth migration

* convert buffer values back to MB on dump/save config

* Revert "that's a big buffer (#585)"

This reverts commit bf490dd1de8d796e33ab6044cbc826b5ca290009.

# Conflicts:
#	src/Daemon/DaemonConfiguration.h

* Let's get back to the basics

* Consistent naming of the variables to reflect that it's a MegaByte (MB)

* Pushing out LWMA update to 1.2M

* Taking the P2P current version back down a notch as there will be no fork for LWMA @ 1M

* Update CryptoNoteCheckpoints.h

* Updated fix_boost_aarch64 directions for c++17 in main repo

* Moved reference to fix_boost_aarch64.sh to main repo in .travis.yml

* Back to one build deployed per platform

* Found some left over garbage in the NetNode files (#593)

* Little bit of fall cleaning never hurt anybody

* If this does nothing then it doesn't need to be here...

* serviceConfig is defined in the header and if it's defined here then it doesn't get passed correctly back up to main.cpp

* Bumping build number for 0.10.0